GreenLight.Loaders (green_light v0.1.0)

This is module that automates the loading of entities based on path parameters.

It is used by the authorization system to enforce access permissions against the loaded entities.

Functions

call(conn, arg)

defloader(key, callback)
(macro)

Create a loader function using common Phoenix patterns. It assumes that the parameter is named id and when used as part of a nested resource setup it will be key_id.

It can thus be used like this:

defloader(:study, &Link.Studies.get_study!/1)

entities(conn)

entity_loader(loader, opts \\ [])
(macro)

init(conf)

load_entities(loaders, conn)